requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.10.14=0
cairo=1.16.0=hf32fb01_1
certifi=2020.6.20=py36_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2021.3.0=pyhd3eb1b0_0
dataclasses=0.8=pyh4f3eec9_6
decorator=5.1.0=pyhd3eb1b0_0
ffmpeg=3.4=h7985aa0_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
glib=2.69.1=h5202010_0
graphite2=1.3.14=h23475e2_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.1=h9caa474_1
icu=58.2=he6710b0_3
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=1.900.1=hd497a04_4
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.4.1=h5b8497f_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.3=h7f8727e_2
networkx=2.5=py_0
ninja=1.10.2=h5e70eb0_2
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
opencv=3.3.1=py36h6cbbc71_1
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1m=h7f8727e_0
pcre=8.45=h295c915_0
pillow=8.3.1=py36h2c7a002_0
pip=21.2.2=py36h06a4308_0
pixman=0.40.0=h7f8727e_1
protobuf=3.19.1=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.1=py3.6_cuda11.0.221_cudnn8.0.5_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.4.1=py36h27cfd23_1
readline=8.1=h27cfd23_0
scikit-image=0.17.2=py36hdf5156a_0
scipy=1.2.1=py36h7c811a0_0
setuptools=58.0.4=py36h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
tensorboardx=1.4=pypi_0
tifffile=2020.10.1=py36hdd07704_2
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.7.2=py36
torchvision=0.8.2=py36_cu110
tornado=6.1=py36h27cfd23_0
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
